More work on the debugger
Moved the code around into the debugger module, the machine.rs file is
long enough as it is. The debugger is theoretically done here and I
added a decent bit of testing for it as well as doing some manual
testing. Seems alright, though probably could use some more test cases.

In the future its probably a good idea to strip out intermediate
whitespace in debug commands, this is currently not working `step    1`